git-svn-id: svn+ssh://nina.grame.fr/home/svn/guidoar@142 2a570e91-86af-4e14-8b94-b0b5b3996bc3

+++ tools/common.cxx
@@ -30,7 +30,7 @@ float floatArg (const char* str, float nofloat)
 }
 
 //_______________________________________________________________________________
-int intArg (const char* str, float noint)  
+int intArg (const char* str, int noint)  
 {
 	int num=0;
 	int n = sscanf(str, "%d", &num);
